Split swizzle functionality into own file
Rename trait ToVector -> SwizzleToVector and make it doc-private
