+++ godot-core/src/builtin/vectors/mod.rs
@@ -5,6 +5,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+// Must be first.
 mod vector_macros;
 
 mod vector2;
@@ -14,6 +15,7 @@ mod vector3i;
 mod vector4;
 mod vector4i;
 mod vector_axis;
+mod vector_swizzle;
 
 pub use vector2::*;
 pub use vector2i::*;
@@ -22,5 +24,6 @@ pub use vector3i::*;
 pub use vector4::*;
 pub use vector4i::*;
 pub use vector_axis::*;
+pub use vector_swizzle::*;
 
 pub use crate::swizzle;
