+++ godot-core/src/builtin/vectors/vector2i.rs
@@ -126,7 +126,7 @@ impl_common_vector_fns!(Vector2i, i32);
 impl_integer_vector_glam_fns!(Vector2i, real);
 impl_integer_vector_component_fns!(Vector2i, real, (x, y));
 impl_vector_operators!(Vector2i, i32, (x, y));
-impl_from_tuple_for_vector2x!(Vector2i, i32);
+impl_swizzle_trait_for_vector2x!(Vector2i, i32);
 
 // SAFETY:
 // This type is represented as `Self` in Godot, so `*mut Self` is sound.
